Refactor modules into core and utils
