+++ src/dh_workspace/__init__.py
@@ -1,11 +1,11 @@
 """dh_workspace package initialization."""
 
-from .placeholder import greet
-from .chessboard import Chessboard
-from .pieces import ChessPiece, Knight, PieceMove, PieceColor
-from .match import Match
-from .config import CONFIG, Config, configure
-from .logger import logger
+from .utils.placeholder import greet
+from .core.chessboard import Chessboard
+from .core.pieces import ChessPiece, Knight, PieceMove, PieceColor
+from .core.match import Match
+from .utils.config import CONFIG, Config, configure
+from .utils.logger import logger
 
 __all__ = [
     "greet",
