+++ src/dh_workspace/core/__init__.py
@@ -1,0 +1,13 @@
+from .chessboard import Chessboard, Piece
+from .match import Match
+from .pieces import ChessPiece, Knight, PieceMove, PieceColor
+
+__all__ = [
+    "Chessboard",
+    "Piece",
+    "ChessPiece",
+    "PieceMove",
+    "PieceColor",
+    "Knight",
+    "Match",
+]
