+++ src/dh_workspace/placeholder.py
@@ -1,6 +1,1 @@
-"""Placeholder module for the dh_workspace package."""
-
-
-def greet(name: str = "World") -> str:
-    """Return a greeting message."""
-    return f"Hello, {name}!"
+from .utils.placeholder import *
