+++ src/dh_workspace/utils/logger.py
@@ -1,0 +1,9 @@
+import logging
+
+logger = logging.getLogger("dh_workspace")
+_handler = logging.StreamHandler()
+_formatter = logging.Formatter("[%(levelname)s] %(message)s")
+_handler.setFormatter(_formatter)
+logger.addHandler(_handler)
+logger.setLevel(logging.INFO)
+logger.propagate = False
