+++ src/dh_workspace/utils/placeholder.py
@@ -1,0 +1,6 @@
+"""Placeholder module for the dh_workspace package."""
+
+
+def greet(name: str = "World") -> str:
+    """Return a greeting message."""
+    return f"Hello, {name}!"
